Update 2_mapping_vRNAs.sh
A few modifications. The script now can take a -n flag for the input number of reads in the fastq file. Outputs read lengths between 15-30nt will still be outputted as 0 even when no reads of that size exist. If the -n flag is used, then the output will be counts per million.